Update Array_diff.py
previous code failed if a repeated values was in list b e.g. a=[13.5], b=[5,5] remove statement prodcued an error when second value found in b because this value had already been removed from a.  Used while and in to remove value until all are deleted from a, while loop does not execute if value already removed from a